+++ SpeciesDatabaseApi/Mr/Enumerations.cs
@@ -1,0 +1,19 @@
+﻿namespace SpeciesDatabaseApi.Mr;
+
+public enum MrRelationDirection
+{
+    Upper, 
+    Lower,
+    Both
+}
+
+public enum MrRelationType
+{
+    PartOf,
+    PartlyPartOf, 
+    AdjacentTo, 
+    SimilarTo, 
+    AdministrativePartOf, 
+    InfluencedBy,
+    All
+}